requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-colorfield==0.4.2
django-debug-toolbar==3.2.1
django-import-export==2.5.0
django-tinymce==3.3.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
